Added Visual C++ 2013 as officially supported compiler into CMakeLists.txt, fixed some size_t to qint32 implicit conversion warnings

+++ core/src/client/types/Note.cpp
@@ -294,7 +294,7 @@ void Note::tagGuids(QStringList & guids) const
     }
 
     const QList<QString> & localTagGuids = d->m_qecNote.tagGuids;
-    size_t numTagGuids = localTagGuids.size();
+    int numTagGuids = localTagGuids.size();
     if (numTagGuids == 0) {
         QNDEBUG("Note::tagGuids: no tags");
         return;
